Moved files of layout.export.tests package to layout.export package.
