+++ tests/hu.elte.txtuml.layout.export.tests/src/hu/elte/txtuml/layout/export/AllTests.java
@@ -1,4 +1,4 @@
-package hu.elte.txtuml.layout.export.tests;
+package hu.elte.txtuml.layout.export;
 
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
