+++ tests/hu.elte.txtuml.layout.export.tests/src/hu/elte/txtuml/layout/export/Test1.java
@@ -1,4 +1,4 @@
-package hu.elte.txtuml.layout.export.tests;
+package hu.elte.txtuml.layout.export;
 
 import hu.elte.txtuml.api.layout.Alignment;
 import hu.elte.txtuml.api.layout.AlignmentType;
@@ -15,20 +15,20 @@ import hu.elte.txtuml.api.layout.TopMost;
 import hu.elte.txtuml.api.model.Association;
 import hu.elte.txtuml.api.model.Model;
 import hu.elte.txtuml.api.model.ModelClass;
-import hu.elte.txtuml.layout.export.tests.Diagram1.L;
-import hu.elte.txtuml.layout.export.tests.Diagram1.LG1;
-import hu.elte.txtuml.layout.export.tests.Diagram1.MyClass;
-import hu.elte.txtuml.layout.export.tests.Diagram1.NG;
-import hu.elte.txtuml.layout.export.tests.Diagram1.NG3;
-import hu.elte.txtuml.layout.export.tests.Diagram1.NG52;
-import hu.elte.txtuml.layout.export.tests.Diagram1.NGA;
-import hu.elte.txtuml.layout.export.tests.Diagram1.NGB;
-import hu.elte.txtuml.layout.export.tests.Diagram1.NGC;
-import hu.elte.txtuml.layout.export.tests.Model1.A;
-import hu.elte.txtuml.layout.export.tests.Model1.B;
-import hu.elte.txtuml.layout.export.tests.Model1.C;
-import hu.elte.txtuml.layout.export.tests.Model1.D;
-import hu.elte.txtuml.layout.export.tests.Model1.E;
+import hu.elte.txtuml.layout.export.Diagram1.L;
+import hu.elte.txtuml.layout.export.Diagram1.LG1;
+import hu.elte.txtuml.layout.export.Diagram1.MyClass;
+import hu.elte.txtuml.layout.export.Diagram1.NG;
+import hu.elte.txtuml.layout.export.Diagram1.NG3;
+import hu.elte.txtuml.layout.export.Diagram1.NG52;
+import hu.elte.txtuml.layout.export.Diagram1.NGA;
+import hu.elte.txtuml.layout.export.Diagram1.NGB;
+import hu.elte.txtuml.layout.export.Diagram1.NGC;
+import hu.elte.txtuml.layout.export.Model1.A;
+import hu.elte.txtuml.layout.export.Model1.B;
+import hu.elte.txtuml.layout.export.Model1.C;
+import hu.elte.txtuml.layout.export.Model1.D;
+import hu.elte.txtuml.layout.export.Model1.E;
 import hu.elte.txtuml.layout.visualizer.statements.StatementType;
 
 import org.junit.Test;
