+++ tests/hu.elte.txtuml.layout.export.tests/src/hu/elte/txtuml/layout/export/Test2.java
@@ -1,4 +1,4 @@
-package hu.elte.txtuml.layout.export.tests;
+package hu.elte.txtuml.layout.export;
 
 import hu.elte.txtuml.api.layout.Alignment;
 import hu.elte.txtuml.api.layout.AlignmentType;
@@ -15,29 +15,29 @@ import hu.elte.txtuml.api.model.Association;
 import hu.elte.txtuml.api.model.Model;
 import hu.elte.txtuml.api.model.ModelClass;
 import hu.elte.txtuml.layout.export.DiagramType;
-import hu.elte.txtuml.layout.export.tests.Model2.BaseOfColumnTop;
-import hu.elte.txtuml.layout.export.tests.Model2.BaseOfDiamond;
-import hu.elte.txtuml.layout.export.tests.Model2.BetweenBases;
-import hu.elte.txtuml.layout.export.tests.Model2.ColumnBottom;
-import hu.elte.txtuml.layout.export.tests.Model2.ColumnBottom_ColumnBottom;
-import hu.elte.txtuml.layout.export.tests.Model2.ColumnMiddle;
-import hu.elte.txtuml.layout.export.tests.Model2.ColumnMiddle_ColumnBottom;
-import hu.elte.txtuml.layout.export.tests.Model2.ColumnTop;
-import hu.elte.txtuml.layout.export.tests.Model2.ColumnTop_ColumnMiddle;
-import hu.elte.txtuml.layout.export.tests.Model2.DiamondBottom;
-import hu.elte.txtuml.layout.export.tests.Model2.DiamondBottom_DiamondLeft;
-import hu.elte.txtuml.layout.export.tests.Model2.DiamondBottom_DiamondRight;
-import hu.elte.txtuml.layout.export.tests.Model2.DiamondLeft;
-import hu.elte.txtuml.layout.export.tests.Model2.DiamondLeft_DiamondRight;
-import hu.elte.txtuml.layout.export.tests.Model2.DiamondRight;
-import hu.elte.txtuml.layout.export.tests.Model2.DiamondTop;
-import hu.elte.txtuml.layout.export.tests.Model2.DiamondTop_DiamondBottom;
-import hu.elte.txtuml.layout.export.tests.Model2.DiamondTop_DiamondLeft;
-import hu.elte.txtuml.layout.export.tests.Model2.DiamondTop_DiamondRight;
-import hu.elte.txtuml.layout.export.tests.Model2.Random;
-import hu.elte.txtuml.layout.export.tests.Model2.Random_BaseOfColumnTop;
-import hu.elte.txtuml.layout.export.tests.Model2.Random_BaseOfDiamond;
-import hu.elte.txtuml.layout.export.tests.Model2.Random_BetweenBases;
+import hu.elte.txtuml.layout.export.Model2.BaseOfColumnTop;
+import hu.elte.txtuml.layout.export.Model2.BaseOfDiamond;
+import hu.elte.txtuml.layout.export.Model2.BetweenBases;
+import hu.elte.txtuml.layout.export.Model2.ColumnBottom;
+import hu.elte.txtuml.layout.export.Model2.ColumnBottom_ColumnBottom;
+import hu.elte.txtuml.layout.export.Model2.ColumnMiddle;
+import hu.elte.txtuml.layout.export.Model2.ColumnMiddle_ColumnBottom;
+import hu.elte.txtuml.layout.export.Model2.ColumnTop;
+import hu.elte.txtuml.layout.export.Model2.ColumnTop_ColumnMiddle;
+import hu.elte.txtuml.layout.export.Model2.DiamondBottom;
+import hu.elte.txtuml.layout.export.Model2.DiamondBottom_DiamondLeft;
+import hu.elte.txtuml.layout.export.Model2.DiamondBottom_DiamondRight;
+import hu.elte.txtuml.layout.export.Model2.DiamondLeft;
+import hu.elte.txtuml.layout.export.Model2.DiamondLeft_DiamondRight;
+import hu.elte.txtuml.layout.export.Model2.DiamondRight;
+import hu.elte.txtuml.layout.export.Model2.DiamondTop;
+import hu.elte.txtuml.layout.export.Model2.DiamondTop_DiamondBottom;
+import hu.elte.txtuml.layout.export.Model2.DiamondTop_DiamondLeft;
+import hu.elte.txtuml.layout.export.Model2.DiamondTop_DiamondRight;
+import hu.elte.txtuml.layout.export.Model2.Random;
+import hu.elte.txtuml.layout.export.Model2.Random_BaseOfColumnTop;
+import hu.elte.txtuml.layout.export.Model2.Random_BaseOfDiamond;
+import hu.elte.txtuml.layout.export.Model2.Random_BetweenBases;
 import hu.elte.txtuml.layout.visualizer.statements.StatementType;
 
 import org.junit.Test;
